refactor(frontend): 拆分信息中心与自动刷新服务
