+++ lib/services/auto_refresh_service.dart
@@ -26,6 +26,10 @@ import 'college_news_service.dart';
 import 'notification_service.dart';
 import 'wechat_article_service.dart';
 
+part 'auto_refresh_service_init.dart';
+part 'auto_refresh_service_fetch.dart';
+part 'auto_refresh_service_timers.dart';
+
 /// 自动刷新服务（单例）
 /// 根据各渠道配置的间隔定时抓取新消息，发现新消息时推送系统通知
 class AutoRefreshService {
@@ -145,864 +149,30 @@ class AutoRefreshService {
 
   /// 初始化并启动自动刷新
   /// 应在 app 启动时调用一次
-  Future<void> init() async {
-    if (_initialized) return;
-    _initialized = true;
-
-    await _stateService.init();
-    await _notificationService.init();
-
-    // 根据各渠道配置启动对应定时器
-    await _setupTimer(
-      channelKey: 'latestInfo',
-      getInterval: _stateService.getLatestInfoInterval,
-      isEnabled: _stateService.isLatestInfoEnabled,
-      fetchMessages: (knownMessageIds) => _newsService.fetchLatestInfo(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    await _setupTimer(
-      channelKey: 'notice',
-      getInterval: _stateService.getNoticeInterval,
-      isEnabled: _stateService.isNoticeEnabled,
-      fetchMessages: (knownMessageIds) => _newsService.fetchNotices(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 教务处教学动态
-    await _setupTimer(
-      channelKey: 'jwcTeaching',
-      getInterval: () => _stateService.getChannelInterval('jwc'),
-      isEnabled: () => _stateService.isChannelEnabled('jwc'),
-      fetchMessages: (knownMessageIds) => _jwcService.fetchTeachingNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 教务处学生专栏
-    await _setupTimer(
-      channelKey: 'jwcStudent',
-      getInterval: () => _stateService.getChannelInterval('jwc'),
-      isEnabled: () => _stateService.isChannelEnabled('jwc'),
-      fetchMessages: (knownMessageIds) => _jwcService.fetchStudentNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 教务处教师专栏
-    await _setupTimer(
-      channelKey: 'jwcTeacher',
-      getInterval: () => _stateService.getChannelInterval('jwc'),
-      isEnabled: () => _stateService.isChannelEnabled('jwc'),
-      fetchMessages: (knownMessageIds) => _jwcService.fetchTeacherNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 信息技术中心
-    await _setupTimer(
-      channelKey: 'itc',
-      getInterval: () => _stateService.getChannelInterval('itc'),
-      isEnabled: () => _stateService.isChannelEnabled('itc'),
-      fetchMessages: (knownMessageIds) => _itcService.fetchNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 学校官网学校新闻
-    await _setupTimer(
-      channelKey: 'sspuNews',
-      getInterval: () => _stateService.getChannelInterval('sspu_news'),
-      isEnabled: () => _stateService.isChannelEnabled('sspu_news'),
-      fetchMessages: (knownMessageIds) => _officialService.fetchNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 学校官网通知公告
-    await _setupTimer(
-      channelKey: 'sspuNotice',
-      getInterval: () => _stateService.getChannelInterval('sspu_notice'),
-      isEnabled: () => _stateService.isChannelEnabled('sspu_notice'),
-      fetchMessages: (knownMessageIds) => _officialService.fetchNotices(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 学校官网校内活动
-    await _setupTimer(
-      channelKey: 'sspuActivity',
-      getInterval: () => _stateService.getChannelInterval('sspu_activity'),
-      isEnabled: () => _stateService.isChannelEnabled('sspu_activity'),
-      fetchMessages: (knownMessageIds) => _officialService.fetchActivities(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 体育部通知公告
-    await _setupTimer(
-      channelKey: 'sportsNotice',
-      getInterval: () => _stateService.getChannelInterval('sports'),
-      isEnabled: () => _stateService.isChannelEnabled('sports'),
-      fetchMessages: (knownMessageIds) => _sportsService.fetchNotices(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 体育部赛事通知
-    await _setupTimer(
-      channelKey: 'sportsEvent',
-      getInterval: () => _stateService.getChannelInterval('sports'),
-      isEnabled: () => _stateService.isChannelEnabled('sports'),
-      fetchMessages: (knownMessageIds) => _sportsService.fetchEvents(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 保卫处平安动态
-    await _setupTimer(
-      channelKey: 'securityNews',
-      getInterval: () => _stateService.getChannelInterval('security_dept'),
-      isEnabled: () => _stateService.isChannelEnabled('security_dept'),
-      fetchMessages: (knownMessageIds) => _securityService.fetchNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 保卫处宣教专栏
-    await _setupTimer(
-      channelKey: 'securityEducation',
-      getInterval: () => _stateService.getChannelInterval('security_dept'),
-      isEnabled: () => _stateService.isChannelEnabled('security_dept'),
-      fetchMessages: (knownMessageIds) => _securityService.fetchEducation(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 基建处建设要闻
-    await _setupTimer(
-      channelKey: 'constructionNews',
-      getInterval: () => _stateService.getChannelInterval('construction'),
-      isEnabled: () => _stateService.isChannelEnabled('construction'),
-      fetchMessages: (knownMessageIds) => _constructionService.fetchNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 基建处通知公告
-    await _setupTimer(
-      channelKey: 'constructionNotice',
-      getInterval: () => _stateService.getChannelInterval('construction'),
-      isEnabled: () => _stateService.isChannelEnabled('construction'),
-      fetchMessages: (knownMessageIds) => _constructionService.fetchNotices(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 新闻网综合新闻
-    await _setupTimer(
-      channelKey: 'campusNews',
-      getInterval: () => _stateService.getChannelInterval('news_center'),
-      isEnabled: () => _stateService.isChannelEnabled('news_center'),
-      fetchMessages: (knownMessageIds) => _campusService.fetchCampusNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 学生处学工要闻
-    await _setupTimer(
-      channelKey: 'studentNews',
-      getInterval: () => _stateService.getChannelInterval('student_affairs'),
-      isEnabled: () => _stateService.isChannelEnabled('student_affairs'),
-      fetchMessages: (knownMessageIds) => _studentService.fetchNews(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // 学生处通知公告
-    await _setupTimer(
-      channelKey: 'studentNotice',
-      getInterval: () => _stateService.getChannelInterval('student_affairs'),
-      isEnabled: () => _stateService.isChannelEnabled('student_affairs'),
-      fetchMessages: (knownMessageIds) => _studentService.fetchNotices(
-        maxCount: _defaultFetchCount,
-        knownMessageIds: knownMessageIds,
-      ),
-    );
-
-    // ==================== 配置驱动站点渠道 ====================
-    // 学院、中心及职能部门共用 CollegeNewsService，通过 channelId 区分解析配置。
-    final siteChannelIds = [
-      'logistics_center',
-      'foreign_student_office',
-      'intl_exchange_office',
-      'admissions_office',
-      'hr_office',
-      'research_office',
-      'union',
-      'party_org_dept',
-      'united_front_dept',
-      'party_office',
-      'youth_league',
-      'assets_lab_office',
-      'college_cs',
-      'college_im',
-      'college_re',
-      'college_em',
-      'college_ic',
-      'college_imhe',
-      'college_econ',
-      'college_lang',
-      'college_math',
-      'college_art',
-      'college_vte',
-      'college_vt',
-      'college_marx',
-      'college_ce',
-      'center_art_edu',
-      'center_intl',
-      'center_innov',
-      'center_training',
-      'graduate',
-      'lib_center',
-    ];
-    for (final channelId in siteChannelIds) {
-      await _setupTimer(
-        channelKey: channelId,
-        getInterval: () => _stateService.getChannelInterval(channelId),
-        isEnabled: () => _stateService.isChannelEnabled(channelId),
-        fetchMessages: (knownMessageIds) => _collegeService.fetchNews(
-          channelId,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-
-    // 微信公众号渠道（通过公众号平台获取）
-    await _setupTimer(
-      channelKey: 'wechatPublic',
-      getInterval: () => _stateService.getChannelInterval('wechat_public'),
-      isEnabled: () => _stateService.isChannelEnabled('wechat_public'),
-      fetchMessages: (knownMessageIds) async {
-        final maxCount = await _stateService.getChannelAutoFetchCount(
-          'wechat_public',
-          defaultValue: _defaultFetchCount,
-        );
-        return _wechatService.fetchArticles(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        );
-      },
-    );
-
-    // 微信服务号占位 — 未来接入时取消注释
-    // await _setupTimer(channelKey: 'wechatService', ...);
-  }
+  /// 初始化并启动自动刷新
+  /// 应在 app 启动时调用一次
+  Future<void> init() => _initAutoRefreshService(this);
 
   /// 立即抓取所有已启用官网/信息中心渠道的消息并返回合并结果
   /// 用于“刷新官网消息”按钮，不包含微信公众号渠道
-  /// [maxCount] 支持 maxCount 参数的服务使用此值
-  /// :return: 所有已启用官网/信息中心渠道的消息列表
   Future<List<MessageItem>> fetchEnabledSchoolWebsiteMessages({
     int maxCount = 20,
     Future<void> Function(List<MessageItem> messages, int completed, int total)?
     onBatchCompleted,
-  }) async {
-    final futures = <Future<List<MessageItem>>>[];
-    final existingMessages = await _stateService.loadMessages();
-    final knownMessageIds = existingMessages.map((msg) => msg.id).toSet();
-
-    // 信息公开网
-    if (await _stateService.isLatestInfoEnabled()) {
-      futures.add(
-        _newsService.fetchLatestInfo(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isNoticeEnabled()) {
-      futures.add(
-        _newsService.fetchNotices(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-
-    // 职能部门
-    if (await _stateService.isChannelEnabled('jwc')) {
-      futures.add(
-        _jwcService.fetchTeachingNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-      futures.add(
-        _jwcService.fetchStudentNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-      futures.add(
-        _jwcService.fetchTeacherNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('itc')) {
-      futures.add(
-        _itcService.fetchNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('sspu_notice')) {
-      futures.add(
-        _officialService.fetchNotices(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('sspu_news')) {
-      futures.add(
-        _officialService.fetchNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('sspu_activity')) {
-      futures.add(
-        _officialService.fetchActivities(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('sports')) {
-      futures.add(
-        _sportsService.fetchNotices(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-      futures.add(
-        _sportsService.fetchEvents(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('security_dept')) {
-      futures.add(
-        _securityService.fetchNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-      futures.add(
-        _securityService.fetchEducation(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('construction')) {
-      futures.add(
-        _constructionService.fetchNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-      futures.add(
-        _constructionService.fetchNotices(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('news_center')) {
-      futures.add(
-        _campusService.fetchCampusNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-    if (await _stateService.isChannelEnabled('student_affairs')) {
-      futures.add(
-        _studentService.fetchNews(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-      futures.add(
-        _studentService.fetchNotices(
-          maxCount: maxCount,
-          knownMessageIds: knownMessageIds,
-        ),
-      );
-    }
-
-    // 配置驱动站点（教学单位 + 职能部门）
-    const siteChannelIds = [
-      'logistics_center',
-      'foreign_student_office',
-      'intl_exchange_office',
-      'admissions_office',
-      'hr_office',
-      'research_office',
-      'union',
-      'party_org_dept',
-      'united_front_dept',
-      'party_office',
-      'youth_league',
-      'assets_lab_office',
-      'college_cs',
-      'college_im',
-      'college_re',
-      'college_em',
-      'college_ic',
-      'college_imhe',
-      'college_econ',
-      'college_lang',
-      'college_math',
-      'college_art',
-      'college_vte',
-      'college_vt',
-      'college_marx',
-      'college_ce',
-      'center_art_edu',
-      'center_intl',
-      'center_innov',
-      'center_training',
-      'graduate',
-      'lib_center',
-    ];
-    for (final id in siteChannelIds) {
-      if (await _stateService.isChannelEnabled(id)) {
-        futures.add(
-          _collegeService.fetchNews(id, knownMessageIds: knownMessageIds),
-        );
-      }
-    }
-
-    if (futures.isEmpty) return [];
-
-    // 并行执行，单个渠道异常不影响其他渠道
-    var completed = 0;
-    final results = await Future.wait(
-      futures.map((future) async {
-        final messages = await future.catchError((_) => <MessageItem>[]);
-        completed++;
-        await onBatchCompleted?.call(messages, completed, futures.length);
-        return messages;
-      }),
-    );
-    return results.expand((msgs) => msgs).toList();
-  }
-
-  /// 配置并启动单个渠道的定时器
-  /// [channelKey] 渠道标识（用作 Timer map 的 key）
-  /// [getInterval] 获取当前间隔（分钟）的异步方法
-  /// [isEnabled] 检查渠道是否启用的异步方法
-  /// [fetchMessages] 抓取该渠道消息的异步方法
-  Future<void> _setupTimer({
-    required String channelKey,
-    required Future<int> Function() getInterval,
-    required Future<bool> Function() isEnabled,
-    required Future<List<MessageItem>> Function(Set<String> knownMessageIds)
-    fetchMessages,
-  }) async {
-    // 先取消已有定时器
-    _timers[channelKey]?.cancel();
-    _timers.remove(channelKey);
-
-    final enabled = await isEnabled();
-    final intervalMinutes = await getInterval();
-
-    // 渠道未启用或间隔为 0（关闭）则不启动
-    if (!enabled || intervalMinutes <= 0) return;
-
-    final duration = Duration(minutes: intervalMinutes);
-
-    _timers[channelKey] = Timer.periodic(duration, (_) async {
-      await _doRefresh(channelKey, fetchMessages);
-    });
-  }
-
-  /// 执行单次刷新：抓取 → 对比 → 合并持久化 → 推送新消息通知
-  /// [channelKey] 渠道标识（用于日志）
-  /// [fetchMessages] 抓取消息的方法
-  Future<void> _doRefresh(
-    String channelKey,
-    Future<List<MessageItem>> Function(Set<String> knownMessageIds)
-    fetchMessages,
-  ) async {
-    try {
-      // 加载已有消息
-      final existingMessages = await _stateService.loadMessages();
-      final existingIds = existingMessages.map((m) => m.id).toSet();
-
-      // 抓取新消息
-      final fetched = await fetchMessages(existingIds);
-
-      // 找出真正的新消息（ID 不在已有集合中）
-      final newMessages = fetched
-          .where((m) => !existingIds.contains(m.id))
-          .toList();
-
-      if (newMessages.isEmpty) return;
-
-      // 合并并持久化
-      final merged = _stateService.mergeMessages(existingMessages, fetched);
-      await _stateService.saveMessages(merged);
-
-      // 推送系统通知（检查全局开关和勿扰时段）
-      final notifEnabled = await _stateService.isNotificationEnabled();
-      final inDnd = await _stateService.isInDndPeriod();
-      if (notifEnabled && !inDnd && _notificationService.isAvailable) {
-        // 过滤掉单个公众号通知关闭的消息
-        final notifiableMessages = <MessageItem>[];
-        for (final msg in newMessages) {
-          if (msg.mpBookId != null) {
-            final mpEnabled = await _stateService.isMpNotificationEnabled(
-              msg.mpBookId!,
-            );
-            if (!mpEnabled) continue;
-          }
-          notifiableMessages.add(msg);
-        }
-
-        if (notifiableMessages.length == 1) {
-          await _notificationService.show(
-            title: '新消息',
-            body: notifiableMessages.first.title,
-          );
-        } else if (notifiableMessages.length > 1) {
-          await _notificationService.show(
-            title: '${notifiableMessages.length} 条新消息',
-            body: notifiableMessages.take(3).map((m) => m.title).join('\n'),
-          );
-        }
-      }
-    } catch (_) {
-      // 静默失败，下次定时器触发会重试
-    }
-  }
+  }) => _fetchEnabledSchoolWebsiteMessages(
+    this,
+    maxCount: maxCount,
+    onBatchCompleted: onBatchCompleted,
+  );
 
   /// 重新加载某个渠道的定时器配置。
   /// 设置页修改间隔后调用此方法使新间隔生效。
-  /// [channelKey] 可传设置页渠道 ID，也可传内部定时器 key。
-  Future<void> reloadChannel(String channelKey) async {
-    final timerKeys = _timerKeysForChannel(channelKey);
-    if (timerKeys.length != 1 || timerKeys.single != channelKey) {
-      for (final timerKey in timerKeys) {
-        await reloadChannel(timerKey);
-      }
-      return;
-    }
+  Future<void> reloadChannel(String channelKey) =>
+      _reloadAutoRefreshChannel(this, channelKey);
 
-    switch (channelKey) {
-      case 'latestInfo':
-        await _setupTimer(
-          channelKey: 'latestInfo',
-          getInterval: _stateService.getLatestInfoInterval,
-          isEnabled: _stateService.isLatestInfoEnabled,
-          fetchMessages: (knownMessageIds) => _newsService.fetchLatestInfo(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'notice':
-        await _setupTimer(
-          channelKey: 'notice',
-          getInterval: _stateService.getNoticeInterval,
-          isEnabled: _stateService.isNoticeEnabled,
-          fetchMessages: (knownMessageIds) => _newsService.fetchNotices(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'jwcTeaching':
-        await _setupTimer(
-          channelKey: 'jwcTeaching',
-          getInterval: () => _stateService.getChannelInterval('jwc'),
-          isEnabled: () => _stateService.isChannelEnabled('jwc'),
-          fetchMessages: (knownMessageIds) => _jwcService.fetchTeachingNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'jwcStudent':
-        await _setupTimer(
-          channelKey: 'jwcStudent',
-          getInterval: () => _stateService.getChannelInterval('jwc'),
-          isEnabled: () => _stateService.isChannelEnabled('jwc'),
-          fetchMessages: (knownMessageIds) => _jwcService.fetchStudentNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'jwcTeacher':
-        await _setupTimer(
-          channelKey: 'jwcTeacher',
-          getInterval: () => _stateService.getChannelInterval('jwc'),
-          isEnabled: () => _stateService.isChannelEnabled('jwc'),
-          fetchMessages: (knownMessageIds) => _jwcService.fetchTeacherNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'itc':
-        await _setupTimer(
-          channelKey: 'itc',
-          getInterval: () => _stateService.getChannelInterval('itc'),
-          isEnabled: () => _stateService.isChannelEnabled('itc'),
-          fetchMessages: (knownMessageIds) => _itcService.fetchNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'sspuNotice':
-        await _setupTimer(
-          channelKey: 'sspuNotice',
-          getInterval: () => _stateService.getChannelInterval('sspu_notice'),
-          isEnabled: () => _stateService.isChannelEnabled('sspu_notice'),
-          fetchMessages: (knownMessageIds) => _officialService.fetchNotices(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'sspuNews':
-        await _setupTimer(
-          channelKey: 'sspuNews',
-          getInterval: () => _stateService.getChannelInterval('sspu_news'),
-          isEnabled: () => _stateService.isChannelEnabled('sspu_news'),
-          fetchMessages: (knownMessageIds) => _officialService.fetchNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'sspuActivity':
-        await _setupTimer(
-          channelKey: 'sspuActivity',
-          getInterval: () => _stateService.getChannelInterval('sspu_activity'),
-          isEnabled: () => _stateService.isChannelEnabled('sspu_activity'),
-          fetchMessages: (knownMessageIds) => _officialService.fetchActivities(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'sportsNotice':
-        await _setupTimer(
-          channelKey: 'sportsNotice',
-          getInterval: () => _stateService.getChannelInterval('sports'),
-          isEnabled: () => _stateService.isChannelEnabled('sports'),
-          fetchMessages: (knownMessageIds) => _sportsService.fetchNotices(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'sportsEvent':
-        await _setupTimer(
-          channelKey: 'sportsEvent',
-          getInterval: () => _stateService.getChannelInterval('sports'),
-          isEnabled: () => _stateService.isChannelEnabled('sports'),
-          fetchMessages: (knownMessageIds) => _sportsService.fetchEvents(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'securityNews':
-        await _setupTimer(
-          channelKey: 'securityNews',
-          getInterval: () => _stateService.getChannelInterval('security_dept'),
-          isEnabled: () => _stateService.isChannelEnabled('security_dept'),
-          fetchMessages: (knownMessageIds) => _securityService.fetchNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'securityEducation':
-        await _setupTimer(
-          channelKey: 'securityEducation',
-          getInterval: () => _stateService.getChannelInterval('security_dept'),
-          isEnabled: () => _stateService.isChannelEnabled('security_dept'),
-          fetchMessages: (knownMessageIds) => _securityService.fetchEducation(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'constructionNews':
-        await _setupTimer(
-          channelKey: 'constructionNews',
-          getInterval: () => _stateService.getChannelInterval('construction'),
-          isEnabled: () => _stateService.isChannelEnabled('construction'),
-          fetchMessages: (knownMessageIds) => _constructionService.fetchNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'constructionNotice':
-        await _setupTimer(
-          channelKey: 'constructionNotice',
-          getInterval: () => _stateService.getChannelInterval('construction'),
-          isEnabled: () => _stateService.isChannelEnabled('construction'),
-          fetchMessages: (knownMessageIds) => _constructionService.fetchNotices(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'campusNews':
-        await _setupTimer(
-          channelKey: 'campusNews',
-          getInterval: () => _stateService.getChannelInterval('news_center'),
-          isEnabled: () => _stateService.isChannelEnabled('news_center'),
-          fetchMessages: (knownMessageIds) => _campusService.fetchCampusNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'studentNews':
-        await _setupTimer(
-          channelKey: 'studentNews',
-          getInterval: () =>
-              _stateService.getChannelInterval('student_affairs'),
-          isEnabled: () => _stateService.isChannelEnabled('student_affairs'),
-          fetchMessages: (knownMessageIds) => _studentService.fetchNews(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      case 'studentNotice':
-        await _setupTimer(
-          channelKey: 'studentNotice',
-          getInterval: () =>
-              _stateService.getChannelInterval('student_affairs'),
-          isEnabled: () => _stateService.isChannelEnabled('student_affairs'),
-          fetchMessages: (knownMessageIds) => _studentService.fetchNotices(
-            maxCount: _defaultFetchCount,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      // 配置驱动站点渠道（教学单位 / 中心 / 职能部门统一处理）
-      case 'logistics_center':
-      case 'foreign_student_office':
-      case 'intl_exchange_office':
-      case 'admissions_office':
-      case 'hr_office':
-      case 'research_office':
-      case 'union':
-      case 'party_org_dept':
-      case 'united_front_dept':
-      case 'party_office':
-      case 'youth_league':
-      case 'assets_lab_office':
-      case 'college_cs':
-      case 'college_im':
-      case 'college_re':
-      case 'college_em':
-      case 'college_ic':
-      case 'college_imhe':
-      case 'college_econ':
-      case 'college_lang':
-      case 'college_math':
-      case 'college_art':
-      case 'college_vte':
-      case 'college_vt':
-      case 'college_marx':
-      case 'college_ce':
-      case 'center_art_edu':
-      case 'center_intl':
-      case 'center_innov':
-      case 'center_training':
-      case 'graduate':
-      case 'lib_center':
-        await _setupTimer(
-          channelKey: channelKey,
-          getInterval: () => _stateService.getChannelInterval(channelKey),
-          isEnabled: () => _stateService.isChannelEnabled(channelKey),
-          fetchMessages: (knownMessageIds) => _collegeService.fetchNews(
-            channelKey,
-            knownMessageIds: knownMessageIds,
-          ),
-        );
-        break;
-      // 微信公众号渠道
-      case 'wechatPublic':
-        await _setupTimer(
-          channelKey: 'wechatPublic',
-          getInterval: () => _stateService.getChannelInterval('wechat_public'),
-          isEnabled: () => _stateService.isChannelEnabled('wechat_public'),
-          fetchMessages: (knownMessageIds) async {
-            final maxCount = await _stateService.getChannelAutoFetchCount(
-              'wechat_public',
-              defaultValue: _defaultFetchCount,
-            );
-            return _wechatService.fetchArticles(
-              maxCount: maxCount,
-              knownMessageIds: knownMessageIds,
-            );
-          },
-        );
-        break;
-      // 其他渠道占位
-      default:
-        break;
-    }
-  }
+  /// 重新加载所有渠道定时器。
+  Future<void> reloadAll() => _reloadAllAutoRefreshChannels(this);
 
-  /// 重新加载所有渠道定时器
-  /// 适用于应用恢复前台或全局刷新设置后
-  Future<void> reloadAll() async {
-    for (final channelId in _refreshChannelIds) {
-      await reloadChannel(channelId);
-    }
-  }
-
-  /// 销毁所有定时器
   void dispose() {
     for (final timer in _timers.values) {
       timer.cancel();
